pyproject.toml
[project]
name = "complinearity"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apptools>=5.3.1",
    "clip",
    "open-clip-torch>=2.26.1",
    "transformers>=4.43.0",
    "cvxpy>=1.7.1",
    "ecos>=2.0.14",
    "envisage>=7.0.4",
    "ipympl>=0.9.7",
    "ipywidgets>=8.1.7",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "osqp>=1.0.4",
    "pandas>=2.3.0",
    "plotly>=6.2.0",
    "pycddlib-standalone>=3.0.0",
    "pyface>=8.0.0",
    "pynormaliz>=2.21",
    "pyqt5>=5.15.11",
    "pysr>=1.5.9",
    "scikit-image>=0.25.2",
    "pyvista[jupyter]>=0.46.1",
    "pyvistaqt>=0.11.3",
    "panel>=1.4.5",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "scs>=3.2.8",
    "seaborn>=0.13.2",
    "stpyvista>=0.1.4",
    "streamlit>=1.48.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
    "traits>=7.0.2",
    "traitsui>=8.0.0",
    "vtk>=9.5.0",
    "widgetsnbextension>=4.0.14",
    "statsmodels>=0.14.5",
    "gurobipy>=12.0.3",
    "wandb>=0.21.4",
    "h5py>=3.14.0",
    "pytest>=8.4.2",
]

[tool.uv.sources]
clip = { git = "https://github.com/openai/CLIP.git" }

[project.optional-dependencies]
visualization = [
    "mayavi>=4.8.3",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
